[feature] Change interface of reset_task.
**Important**

I'll stop implementing reset_task and reset_list, because it is not good
implementing method to cancel the currently ready tasks. To improve the
processing the redoing canceled and failed task, stop implementing and
close the CLI interface implementing issues. And before the implementing
this features, change the db interface more clearly.

- Currently pending implementing the reset_task and reset_list
  - Will change reset_task processing steps.
  - Currently one-step reset task, but it will change cancel and redo.
  - Change some queries for reset
- Change `db_config` to property function
- Change db connection variable in `mariadb.py` from _mariadb to _db.